fixes to search.js empty() list when searchBtn is clicked. fixes to ytplayer.js validate videoId before playing. fixes to control.js-->deletevideo() when control.videoId video is deleted.

+++ js/modules/control.js
@@ -28,6 +28,12 @@ define(["modules/backEnd"], function(backEnd) {
             var index = control.playList.indexOf(videoId); // video does not exist in list
             if (index < 0)
                 return;
+            if (videoId === control.videoId) {
+                var current = control.playList.indexOf(videoId);
+                var next = (current + 1) % control.playList.length;
+                control.videoId = control.playList[next];
+                control.updateByUi("videoId", control.videoId);
+            }
             control.playList.splice(index, 1);
             backEnd.deleteItem(videoId);
         };
